+++ infinigen/__init__.py
@@ -6,7 +6,7 @@
 import logging
 from pathlib import Path
 
-__version__ = "1.13.2"
+__version__ = "1.14.0-dev"
 
 
 def repo_root():