+++ infinigen/core/placement/animation_policy.py
@@ -322,14 +322,16 @@ class AnimPolicyFollowObject:
         return Vector(pos), None, time, "BEZIER"
 
 
+@gin.configurable
 def validate_keyframe_range(
     obj,
     start_frame,
     end_frame,
     bvhtree,
     validate_pose_func=None,
-    stride=5,  # runs faster but imperfect precision
+    stride=1,  # runs faster but imperfect precision
     check_straight_line=True,  # rules out proposals faster, but has imperfect precision
+    check_keyframe_straight_line=True,
 ):
     last_pos = deepcopy(obj.location)
 
@@ -346,9 +348,10 @@ def validate_keyframe_range(
     for frame_idx in range(start_frame, end_frame + 1, stride):
         bpy.context.scene.frame_set(frame_idx)
 
-        if not freespace_ray_check(last_pos, obj.location):
-            logger.debug(f"{frame_idx=} freespace_ray_check failed")
-            return False
+        if check_keyframe_straight_line:
+            if not freespace_ray_check(last_pos, obj.location):
+                logger.debug(f"{frame_idx=} freespace_ray_check failed")
+                return False
 
         if validate_pose_func is not None and not validate_pose_func(obj):
             # technically we should validate against all cameras, but this would be expensive
@@ -370,6 +373,7 @@ def try_animate_trajectory(
     max_step_tries=50,
     verbose=True,
 ):
+    frame_prev = bpy.context.scene.frame_start
     frame_curr = bpy.context.scene.frame_start
     pbar = tqdm(total=duration_frames) if verbose else None
     while frame_curr < bpy.context.scene.frame_start + duration_frames:
@@ -392,10 +396,15 @@ def try_animate_trajectory(
             step_frames = int(duration * bpy.context.scene.render.fps) + 1
             step_end_frame = frame_curr + step_frames
 
-            keyframe(obj, loc, rot, step_end_frame, interp="BEZIER")
+            keyframe(obj, loc, rot, step_end_frame, interp=interp)
 
             if not validate_keyframe_range(
-                obj, frame_curr, step_end_frame, bvh, validate_pose_func
+                obj,
+                frame_prev,
+                step_end_frame,
+                bvh,
+                validate_pose_func,
+                check_straight_line=False,
             ):
                 logger.debug(
                     f"validate_keyframe_range failed on moving {obj.location} to {loc}"
@@ -409,7 +418,10 @@ def try_animate_trajectory(
 
             if verbose:
                 pbar.update(
-                    min(step_frames, duration_frames - frame_curr)
+                    min(
+                        step_frames,
+                        bpy.context.scene.frame_start + duration_frames - frame_curr,
+                    )
                 )  # dont overshoot the pbar, it makes the formatting not nice
 
             break  # we found a good pose
@@ -417,6 +429,7 @@ def try_animate_trajectory(
         else:  # for-else block triggers when for loop terminates w/o a break statement
             return False
 
+        frame_prev = frame_curr
         frame_curr = step_end_frame
         bpy.context.scene.frame_current = frame_curr
 
@@ -601,10 +614,10 @@ def animate_trajectory(
         if attempt > 0 and retry_rotation:
             obj.rotation_euler.z = U(0, 2 * np.pi)
 
-        if hasattr(policy_func, "reset"):
+        if hasattr(policy_func, "reset") and attempt > 0:
             policy_func.reset()
 
-        keyframe(obj, obj.location, obj.rotation_euler, 0, interp="LINEAR")
+        keyframe(obj, obj.location, obj.rotation_euler, 0)
         try_animate_trajectory_func = (
             try_animate_trajectory
             if not path_finding_enabled
@@ -637,26 +650,26 @@ def animate_trajectory(
                         (
                             obj.animation_data.action.fcurves[0]
                             .keyframe_points[j]
-                            .co.y,
+                            .co.x,
                             obj.animation_data.action.fcurves[1]
                             .keyframe_points[j]
                             .co.y,
                             obj.animation_data.action.fcurves[2]
                             .keyframe_points[j]
-                            .co.y,
+                            .co.z,
                         )
                     )
                     kf_rots.append(
                         (
                             obj.animation_data.action.fcurves[3]
                             .keyframe_points[j]
-                            .co.y,
+                            .co.x,
                             obj.animation_data.action.fcurves[4]
                             .keyframe_points[j]
                             .co.y,
                             obj.animation_data.action.fcurves[5]
                             .keyframe_points[j]
-                            .co.y,
+                            .co.z,
                         )
                     )
                 obj.animation_data_clear()