+++ tests/integration/launch.sh
@@ -1,4 +1,7 @@
+#!/bin/bash
+
 OUTPUT_PATH=$1
+shift
 
 if [ -z "$OUTPUT_PATH" ]; then
     echo "Please provide an output path"
@@ -7,6 +10,53 @@ fi
 
 # make outputs group-writable
 umask 0002
+
+# Initialize arrays to hold the arguments
+arg1_strings=()
+arg2_strings=()
+arg3_strings=()
+
+# Parse the command-line arguments
+while [[ $# -gt 0 ]]; do
+  key="$1"
+
+  case $key in
+    --nature_configs)
+      shift  # Move past the option
+      # Collect arguments until the next option or end of input
+      while [[ $# -gt 0 && $1 != --* ]]; do
+        arg1_strings+=("$1")
+        shift
+      done
+      ;;
+    --indoors_configs)
+      shift  # Move past the option
+      # Collect arguments until the next option or end of input
+      while [[ $# -gt 0 && $1 != --* ]]; do
+        arg2_strings+=("$1")
+        shift
+      done
+      ;;
+    --pipeline_configs)
+      shift  # Move past the option
+      # Collect arguments until the next option or end of input
+      while [[ $# -gt 0 && $1 != --* ]]; do
+        arg3_strings+=("$1")
+        shift
+      done
+      ;;
+    *)
+      echo "Unknown option: $key"
+      exit 1
+      ;;
+  esac
+done
+
+
+# Combine the arrays into single strings
+parsed_nature_configs="${arg1_strings[*]}"
+parsed_indoors_configs="${arg2_strings[*]}"
+parsed_pipeline_configs="${arg3_strings[*]}"
 
 # Environment Variables for Opting In/Out
 RUN_INDOOR=${RUN_INDOOR:-1}
@@ -29,8 +79,8 @@ OUTPUT_PATH=$OUTPUT_PATH/$VERSION_STRING
 if [ "$RUN_INDOOR" -eq 1 ]; then
     for indoor_type in DiningRoom Bathroom Bedroom Kitchen LivingRoom; do
         python -m infinigen.datagen.manage_jobs --output_folder $OUTPUT_PATH/${JOBTAG}_scene_indoor_$indoor_type \
-        --num_scenes 3 --cleanup big_files --configs singleroom.gin fast_solve.gin --overwrite \
-        --pipeline_configs slurm monocular blender_gt indoor_background_configs.gin \
+        --num_scenes 3 --cleanup big_files --configs singleroom.gin fast_solve.gin $parsed_indoors_configs --overwrite \
+        --pipeline_configs slurm.gin monocular.gin blender_gt.gin indoor_background_configs.gin $parsed_pipeline_configs \
         --pipeline_overrides get_cmd.driver_script=infinigen_examples.generate_indoors sample_scene_spec.seed_range=[0,100] slurm_submit_cmd.slurm_nodelist=$NODECONF \
         --overrides compose_indoors.terrain_enabled=True restrict_solving.restrict_parent_rooms=\[\"$indoor_type\"\] compose_indoors.solve_small_enabled=False &
     done
@@ -41,10 +91,18 @@ if [ "$RUN_NATURE" -eq 1 ]; then
     for nature_type in arctic canyon cave coast coral_reef desert forest kelp_forest mountain plain river snowy_mountain under_water; do
         python -m infinigen.datagen.manage_jobs --output_folder $OUTPUT_PATH/${JOBTAG}_scene_nature_$nature_type \
         --num_scenes 3 --cleanup big_files --overwrite \
-        --configs $nature_type.gin \
-        --pipeline_configs slurm monocular blender_gt \
+        --configs $nature_type.gin dev.gin $parsed_nature_configs \
+        --pipeline_configs slurm.gin monocular.gin blender_gt.gin $parsed_pipeline_configs \
         --pipeline_overrides sample_scene_spec.seed_range=[0,100] &
     done
+fi
+
+if [ -n "$parsed_nature_configs" ]; then
+    parsed_nature_configs="--configs $parsed_nature_configs "
+fi
+
+if [ -n "$parsed_indoors_configs" ]; then
+    parsed_indoors_configs="--configs $parsed_indoors_configs "
 fi
 
 # Objects
@@ -52,10 +110,12 @@ if [ "$RUN_OBJECTS" -eq 1 ]; then
 
     python -m infinigen_examples.generate_individual_assets \
     -f tests/assets/list_nature_meshes.txt --output_folder $OUTPUT_PATH/${JOBTAG}_asset_nature_meshes \
+    $parsed_nature_configs \
     --slurm --n_workers 100 -n 3 --gpu & 
 
     python -m infinigen_examples.generate_individual_assets \
     -f tests/assets/list_indoor_meshes.txt --output_folder $OUTPUT_PATH/${JOBTAG}_asset_indoor_meshes \
+    $parsed_indoors_configs \
     --slurm --n_workers 100 -n 3 --gpu &
 fi
 
@@ -63,12 +123,12 @@ fi
 if [ "$RUN_MATERIALS" -eq 1 ]; then
 
     python -m infinigen_examples.generate_individual_assets \
-    -f tests/assets/list_materials.txt --output_folder $OUTPUT_PATH/${JOBTAG}_asset_new_materials \
+    -f tests/assets/list_materials.txt --output_folder $OUTPUT_PATH/${JOBTAG}_asset_new_materials $parsed_indoors_configs \
     --slurm --n_workers 100 -n 3 --gpu & 
 
 
     python -m infinigen_examples.generate_individual_assets \
-    -f tests/assets/list_materials_deprecated_interface.txt --output_folder $OUTPUT_PATH/${JOBTAG}_asset_deprec_materials \
+    -f tests/assets/list_materials_deprecated_interface.txt --output_folder $OUTPUT_PATH/${JOBTAG}_asset_deprec_materials $parsed_nature_configs \
     --slurm --n_workers 100 -n 3 --gpu &
 fi