+++ docs/CHANGELOG.md
@@ -203,3 +203,8 @@ v1.13.1
 
 v1.13.2
 - Bugfix invalid Brick kwargs and `1-2 args execution context is supported` during indoor room generation
+
+
+v1.14.0
+- Add option to densely subdivide room meshes using OcMesher
+- Add RRT-Star camera pathfinding for video viewpoint animation